Add files via upload
As in joined in web development course to learn HTML , CSS , MongoDB ,NODE JS , Express JS , Java Script. So after completion of my in carizo platform , they made a team and give major project to develop website. On that basis  we developed these website. I did HTMl and CSS for the website.